CA/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Data / Excel
pandas==2.2.2
openpyxl==3.1.5

# PDF & OCR (Invoice Extraction)
pymupdf==1.27.1
pytesseract==0.3.10
Pillow==12.1.1

# HTTP & API
requests==2.32.5
httpx==0.27.0

# Validation
pydantic[email]

# Supabase
supabase==2.4.4
postgrest==0.16.4

# Auth & Security
bcrypt==4.1.3
python-jose==3.3.0
passlib==1.7.4

# Environment & Config
python-dotenv==1.0.1

# Email (Gmail SMTP)
email-validator==2.1.1

# Utils / Timezone
pytz==2024.1

# Async support (FastAPI internal)
anyio==4.4.0

# Machine learning
scikit-learn
